+++ libft/libft.h
@@ -72,8 +72,8 @@ void				ft_lstclear(t_list **lst, void (*del)(void*));
 void				ft_lstiter(t_list *lst, void (*f)(void *));
 t_list				*ft_lstmap(t_list *lst, void *(*f)(void *),
 					void (*del)(void *));
-int					ft_get_size_x(unsigned long int n);
-char				*ft_itoa_x(unsigned long int n);
+int					ft_get_size_x(unsigned int n);
+char				*ft_itoa_x(unsigned int n);
 void				ft_str_toupper(char *str);
 char				*ft_itoa_ll(long long int n);
